requirements.txt
pydantic>=2.0.0
pydantic-ai[google]>=0.1.0
pydantic-ai[anthropic]>=0.1.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
streamlit>=1.30.0
nest-asyncio>=1.6.0
sniffio>=1.3.0
requests>=2.31.0
networkx>=3.2
sentence-transformers>=2.2.0
pinecone-client>=3.0.0
